+++ benches/src/clip.rs
@@ -19,11 +19,11 @@ fn do_clip_tiny_skia(aa: bool, bencher: &mut Bencher) {
 
         let clip_path = clip_path.transform(Transform::from_row(1.0, -0.5, 0.0, 1.0, 0.0, 300.0)).unwrap();
 
-        let mut clip_mask = ClipMask::new();
-        clip_mask.set_path(1000, 1000, &clip_path, FillRule::EvenOdd, aa);
+        let mut mask = Mask::new();
+        mask.set_path(1000, 1000, &clip_path, FillRule::EvenOdd, aa);
 
         // Do not use fill_rect, because it is very slow by itself.
-        pixmap.fill_path(&path, &paint, FillRule::Winding, Transform::identity(), Some(&clip_mask));
+        pixmap.fill_path(&path, &paint, FillRule::Winding, Transform::identity(), Some(&mask));
     });
 }